Accidently pushed before actuall commit. Now the changes should have been made:
I have now removed cases where the ++ operator was "not needed", since we did not append anything to the end of a list.
The operator is however still used in the code but only where we actually append to a list.

For example the code example for reverse is still:

  def reverse([]), do: []
  def reverse([head | tail]), do: reverse(tail) ++ [head]

Would you like me to optimize this code regarding performance?
Hence refactor the code to remove the slow append command.